requirements.txt
asgiref==3.7.2
Django==4.0.3
djangorestframework==3.14.0
future==0.18.3
myapp==0.1.dev0
mysqlclient==2.2.1
PyMySQL==1.1.0
pytz==2023.3.post1
serializers==0.2.4
sqlparse==0.4.4
typing_extensions==4.9.0
